first version

+++ src/python/WMCore/Storage/DeleteMgr.py
@@ -21,8 +21,9 @@ from WMCore.Storage.StageOutError import StageOutInitError
 from WMCore.WMException import WMException
 
 
-# from WMCore.Storage.SiteLocalConfig import loadSiteLocalConfig
+from WMCore.Storage.SiteLocalConfig import stageOutStr
 
+from WMCore.Storage.RucioFileCatalog import storageJsonPath,readRFC 
 
 class DeleteMgrError(WMException):
     """
@@ -47,29 +48,29 @@ class DeleteMgr(object):
     """
 
     def __init__(self, **overrideParams):
+        
         self.override = False
         self.logger = overrideParams.pop("logger", logging.getLogger())
-        self.overrideConf = overrideParams
+        self.overrideParams = overrideParams
         if overrideParams != {}:
             self.override = True
 
-        #  //
-        # // Try an get the TFC for the site
-        # //
-        self.tfc = None
+        self.bypassImpl = False
 
-        from WMCore.Storage.SiteLocalConfig import loadSiteLocalConfig
+        self.stageOuts_rfcs = [] #pairs of stageOut and Rucio file catalog
+
+        self.overrideConf = None #will be initialized later in initialiseOverride
 
         self.numberOfRetries = 3
         self.retryPauseTime = 600
-        self.pnn = None
-        self.fallbacks = []
+
+        from WMCore.Storage.SiteLocalConfig import loadSiteLocalConfig
 
         #  //
         # // If override isnt None, we dont need SiteCfg, if it is
         # //  then we need siteCfg otherwise we are dead.
 
-        if self.override == False:
+        if not self.override:
             self.siteCfg = loadSiteLocalConfig()
 
         if self.override:
@@ -84,43 +85,52 @@ class DeleteMgr(object):
         Extract required information from site conf and TFC
 
         """
+        
+        self.stageOuts = self.siteCfg.stageOuts
 
-        implName = self.siteCfg.localStageOut.get("command", None)
-        if implName == None:
-            msg = "Unable to retrieve local stage out command\n"
-            msg += "From site config file.\n"
-            msg += "Unable to perform StageOut operation"
-            raise StageOutInitError(msg)
-        msg = "Local Stage Out Implementation to be used is:"
-        msg += "%s\n" % implName
+        msg = ""
+        msg += "There are %s stage out definitions.\n" % len(self.stageOuts)
+        
+        for stageOut in self.stageOuts:
+            msg = ""
+            for k in ['phedex-node','command','storageSite','volume','protocol']:
+                v = stageOut.get(k)
+                if v is None:
+                    msg += "Unable to retrieve "+k+" of this stageOut: \n"
+                    msg += stageOutStr(stageOut) + "\n"
+                    msg += "From site config file.\n"
+                    msg += "Continue to the next stageOut.\n"
+                    logging.info(msg)
+                    continue
 
-        pnn = self.siteCfg.localStageOut.get("phedex-node", None)
-        if pnn == None:
-            msg = "Unable to retrieve local stage out phedex-node\n"
-            msg += "From site config file.\n"
-            msg += "Unable to perform StageOut operation"
-            raise StageOutInitError(msg)
-        msg += "Local Stage Out PNN to be used is %s\n" % pnn
-        catalog = self.siteCfg.localStageOut.get("catalog", None)
-        if catalog == None:
-            msg = "Unable to retrieve local stage out catalog\n"
-            msg += "From site config file.\n"
-            msg += "Unable to perform StageOut operation"
-            raise StageOutInitError(msg)
-        msg += "Local Stage Out Catalog to be used is %s\n" % catalog
+            storageSite = stageOut.get("storageSite")
+            volume = stageOut.get("volume")
+            protocol = stageOut.get("protocol")
+            command = stageOut.get("command")
+            pnn = stageOut.get("phedex-node")
+            
+            msg += "\tStage out to : %s using: %s \n" % (pnn, command)
+            
+            try:
+                aPath = storageJsonPath(self.siteCfg.siteName,self.siteCfg.subSiteName,storageSite)
+                rfc = readRFC(aPath,storageSite,volume,protocol)
+                self.stageOuts_rfcs.append((stageOut,rfc))
+                msg += "Rucio File Catalog has been loaded:\n"
+                msg += str(self.stageOuts_rfcs[-1][1])
+            except Exception as ex:
+                msg += "Unable to load Rucio File Catalog:\n"
+                msg += "This stage out will not be attempted:\n"
+                msg += stageOutStr(stageOut) + '\n'
+                msg += str(ex)
+                logging.info(msg)
+                continue
 
-        try:
-            self.tfc = self.siteCfg.trivialFileCatalog()
-            msg += "Trivial File Catalog has been loaded:\n"
-            msg += str(self.tfc)
-        except Exception as ex:
-            msg = "Unable to load Trivial File Catalog:\n"
-            msg += "Local stage out will not be attempted\n"
-            msg += str(ex)
+        #no Rucio file catalog is initialized
+        if not self.stageOuts_rfcs:
             raise StageOutInitError(msg)
 
         self.logger.info(msg)
-        self.pnn = pnn
+        
         return
 
     def initialiseOverride(self):
@@ -130,8 +140,7 @@ class DeleteMgr(object):
         Extract and verify that the Override parameters are all present
 
         """
-        overrideConf = self.overrideConf
-        overrideParams = {
+        self.overrideConf = {
             "command": None,
             "option": None,
             "phedex-node": None,
@@ -139,26 +148,26 @@ class DeleteMgr(object):
         }
 
         try:
-            overrideParams['command'] = overrideConf['command']
-            overrideParams['phedex-node'] = overrideConf['phedex-node']
-            overrideParams['lfn-prefix'] = overrideConf['lfn-prefix']
+            self.overrideConf['command'] = self.overrideParams['command']
+            self.overrideConf['phedex-node'] = self.overrideParams['phedex-node']
+            self.overrideConf['lfn-prefix'] = self.overrideParams['lfn-prefix']
         except Exception as ex:
-            msg = "Unable to extract Override parameters from config:\n"
-            msg += str(overrideConf)
+            msg = "Unable to extract override parameters from config:\n"
+            msg += str(ex)
             raise StageOutInitError(msg)
-        if 'option' in overrideConf:
-            if len(overrideConf['option']) > 0:
-                overrideParams['option'] = overrideConf['option']
+        if 'option' in self.overrideParams:
+            if len(self.overrideParams['option']) > 0:
+                self.overrideConf['option'] = self.overrideParams['option']
             else:
-                overrideParams['option'] = ""
+                self.overrideConf['option'] = ""
 
         msg = "=======Delete Override Initialised:================\n"
-        for key, val in viewitems(overrideParams):
+        for key, val in viewitems(self.overrideConf):
             msg += " %s : %s\n" % (key, val)
         msg += "=====================================================\n"
+
         self.logger.info(msg)
-        self.fallbacks.append(overrideParams)
-        self.pnn = overrideParams['phedex-node']
+        
         return
 
     def __call__(self, fileToDelete):
@@ -171,34 +180,28 @@ class DeleteMgr(object):
         self.logger.info("==>Working on file: %s" % fileToDelete['LFN'])
 
         lfn = fileToDelete['LFN']
-        fileToDelete['PNN'] = self.pnn
 
         deleteSuccess = False
-
-        #  //
-        # // No override => use local-stage-out from site conf
-        # //  invoke for all files and check failures/successes
+        
         if not self.override:
-            self.logger.info("===> Attempting To Delete.")
+            logging.info("===> Attempting to delete with %s stage outs", len(self.stageOuts))
+            for stageOut_rfc in self.stageOuts_rfcs:
+                if not deleteSuccess:
+                    try:
+                        fileToDelete['PNN'] = stageOut_rfc[0]['phedex-node']
+                        fileToDelete['PFN'] = self.deleteLFN(lfn, stageOut_rfc)
+                        deleteSuccess = True
+                        break
+                    except Exception as ex:
+                        continue
+        else:
+            logging.info("===> Attempting stage outs from override")
             try:
-                fileToDelete['PFN'] = self.deleteLFN(lfn)
+                fileToDelete['PNN'] = self.overrideConf['phedex-node']
+                fileToDelete['PFN'] = self.deleteLFN_override(lfn)
                 deleteSuccess = True
             except Exception as ex:
                 self.logger.error("===> Local file deletion failure. Exception:\n%s", str(ex))
-
-        if not deleteSuccess and len(self.fallbacks) > 0:
-            #  //
-            # // Still here => override start using the fallback stage outs
-            # //  If override is set, then that will be the only fallback available
-            self.logger.info("===> Attempting To Delete files with fallback.")
-            for fallback in self.fallbacks:
-                if not deleteSuccess:
-                    try:
-                        fileToDelete['PFN'] = self.deleteLFN(lfn, fallback)
-                        deleteSuccess = True
-                    except Exception as ex:
-                        continue
-
         if deleteSuccess:
             msg = "===> Delete Successful:\n"
             msg += "====> LFN: %s\n" % fileToDelete['LFN']
@@ -210,30 +213,46 @@ class DeleteMgr(object):
             msg = "Unable to delete file:\n"
             msg += fileToDelete['LFN']
             raise StageOutFailure(msg, **fileToDelete)
-
-    def deleteLFN(self, lfn, override=None):
+    
+    def deleteLFN(self, lfn, stageOut_rfc):
         """
         deleteLFN
 
-        Given the lfn and local stage out params, invoke the delete
+        Given the lfn and an stageOut Rucio file catalog pair, invoke the delete
 
-        if override is used the follwoing params should be defined:
+        """
+        from WMCore.Storage.StageOutMgr import searchRFC 
+        command = stageOut_rfc[0]['command']
+        pfn = searchRFC(stageOut_rfc[1],lfn)
+
+        if pfn == None:
+            msg = "Unable to match lfn to pfn: \n  %s" % lfn
+            raise StageOutFailure(msg, LFN=lfn, STAGEOUT=stageOutStr(stageOut_rfc[0]))
+
+        return self.deletePFN(pfn, lfn, command)
+    
+
+    def deleteLFN_override(self, lfn):
+        """
+        deleteLFN_override
+
+        Given the lfn invoke the delete using override config
+
+        the follwoing params should be defined for override
         command - the stage out impl plugin name to be used
         option - the option values to be passed to that command (None is allowed)
         lfn-prefix - the LFN prefix to generate the PFN
         phedex-node - the Name of the PNN to which the file is being xferred
         """
 
-        if override:
-            command = override['command']
-            pfn = "%s%s" % (override['lfn-prefix'], lfn)
-        else:
-            command = self.siteCfg.localStageOut['command']
-            pfn = self.searchTFC(lfn)
+        command = self.overrideConf['command']
+        pfn = None
+        if self.overrideConf['lfn-prefix'] is not None:
+            pfn = "%s%s" % (self.overrideConf['lfn-prefix'], lfn)
 
-        if pfn == None:
-            msg = "Unable to match lfn to pfn: \n  %s" % lfn
-            raise StageOutFailure(msg, LFN=lfn, TFC=str(self.tfc))
+        if pfn is None:
+            msg = "Unable to match lfn to pfn using lfn-prefix: \n %s" % lfn
+            raise StageOutFailure(msg, LFN=lfn, LFNPREFIX=self.overrideConf['lfn-prefix'])
 
         return self.deletePFN(pfn, lfn, command)
 
@@ -253,68 +272,11 @@ class DeleteMgr(object):
         impl.retryPause = self.retryPauseTime
 
         try:
-            impl.removeFile(pfn)
+            if not self.bypassImpl:
+                impl.removeFile(pfn)
         except Exception as ex:
             self.logger.error("Failed to delete file: %s", pfn)
             ex.addInfo(Protocol=command, LFN=lfn, TargetPFN=pfn)
             raise ex
 
         return pfn
-
-    #    def reportStageOutFailure(self, stageOutExcep):
-    #        """
-    #        _reportStageOutFailure_
-    #
-    #        When a stage out failure occurs, report it to the input
-    #        framework job report.
-    #
-    #        - *stageOutExcep* : Instance of on of the StageOutError derived classes
-    #
-    #        """
-    #        errStatus = stageOutExcep.data["ErrorCode"]
-    #        errType = stageOutExcep.data["ErrorType"]
-    #        desc = stageOutExcep.message
-    #
-    #        errReport = self.inputReport.addError(errStatus, errType)
-    #        errReport['Description'] = desc
-    #        return
-
-    def searchTFC(self, lfn):
-        """
-        _searchTFC_
-
-        Search the Trivial File Catalog for the lfn provided,
-        if a match is made, return the matched PFN
-
-        """
-        if self.tfc == None:
-            msg = "Trivial File Catalog not available to match LFN:\n"
-            msg += lfn
-            self.logger.error(msg)
-            return None
-        if self.tfc.preferredProtocol == None:
-            msg = "Trivial File Catalog does not have a preferred protocol\n"
-            msg += "which prevents local stage out for:\n"
-            msg += lfn
-            self.logger.error(msg)
-            return None
-
-        pfn = self.tfc.matchLFN(self.tfc.preferredProtocol, lfn)
-        if pfn == None:
-            msg = "Unable to map LFN to PFN:\n"
-            msg += "LFN: %s\n" % lfn
-            return None
-
-        msg = "LFN to PFN match made:\n"
-        msg += "LFN: %s\nPFN: %s\n" % (lfn, pfn)
-        self.logger.info(msg)
-        return pfn
-
-##if __name__ == '__main__':
-##    import StageOut.Impl
-##    mgr = StageOutMgr()
-##    pfn = "/home/evansde/work/PRODAGENT/work/JobCreator/RelValMinBias-170pre12/Processing/RelValMinBias-170pre12-Processing.tar.gz"
-##    lfn = "/store/unmerged/mc/2007/11/13/RelVal-RelValMinBias-1194987281/GEN-SIM-DIGI-RECO/0201/DCCP-FNAL-TEST.dat"
-
-##    mgr.searchTFC(lfn)
-##    mgr(LFN = lfn, PFN = pfn, GUID=None)
